backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clerk-backend-api>=3.0.3",
    "fastapi>=0.115.13",
    "google-genai>=1.21.1",
    "openai>=1.88.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "svix>=1.56.0",
    "uvicorn>=0.34.3",
]
